archie-ai-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.0.0
pyjwt==2.8.0
cryptography==41.0.7
google-cloud-speech==2.21.0
google-generativeai==0.3.2
python-json-logger==2.0.7
pydantic==2.5.0
python-jose[cryptography]==3.3.0 